Fixed Turrets/Bullets

+++ Model/Turret.cs
@@ -51,7 +51,7 @@ namespace SurfaceTower.Model
     }
     public void Shoot()
     {
-      Bullet bullet = new Bullet(position.X, position.Y, orientation, 5, strength, effect);
+      Bullet bullet = new Bullet((int)position.X, (int)position.Y, orientation, 5, strength, effect);
       BaseModel.INSTANCE.Bullets.Add(bullet);
     }
     #endregion
